edit table

+++ cdn_client/database/migrations/2022_09_22_081917_create_role_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name', 100)->unique()->comment("唯一名稱");
             $table->string('key', 150)->unique()->comment("唯一key");
             $table->boolean('status')->comment("狀態(開,關)")->default(1);
-            $table->integer('weight')->nullable()->comment("權重(優先順序 重=高)");
+            $table->smallInteger('weight')->nullable()->comment("權重(優先順序 重=高)");
             $table->string('remark', 5000)->nullable()->comment("備註");
             $table->timestamps();
         });
